pyproject.toml
[tool.poetry]
name = "detection-dataset"
version = "0.1.0"
description = "Easily convert datasets between different formats for object detection"
authors = ["Jerome Blin <jerome@grover.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/blinjrm/detection-dataset"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [{ include = "detection_dataset", from = "src" }]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.2"
"ruamel.yaml" = "^0.17.21"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
pre-commit = "^2.18.1"
mypy = "^0.950"
pytest-cov = "^3.0.0"
jupyterlab = "^3.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line_length = 120
target-version = ['py38']
